+++ mrcrowbar/models.py
@@ -20,5 +20,5 @@ from mrcrowbar.fields import ParseError, FieldValidationError, Field, \
 from mrcrowbar.blocks import Block, Unknown
 from mrcrowbar.loaders import Loader
 from mrcrowbar.transforms import Transform, TransformResult
-from mrcrowbar.checks import CheckException, Check, Const, Updater
+from mrcrowbar.checks import CheckException, Check, Const, Pointer, Updater
 from mrcrowbar.views import View, Store, LinearStore, StoreRef